drivers: Add pressure measurement for sensors/bmp280

+++ src/drivers/sensors/bmp280/bmp280.h
@@ -12,8 +12,8 @@
 #include <stdint.h>
 
 extern int bmp280_init(void);
-extern int16_t bmp280_get_temperature(void);
-extern int16_t bmp280_get_humidity(void);
-extern int16_t bmp280_get_pressure(void);
+extern int32_t bmp280_get_temperature(void);
+extern int32_t bmp280_get_humidity(void);
+extern int32_t bmp280_get_pressure(void);
 
 #endif /* BMP280_H_ */
